main: refactor wifi files
